+++ src/test/kotlin/me/bossm0n5t3r/leetcode/binaryTreeInorderTraversal/BinaryTreeInorderTraversalTest.kt
@@ -1,13 +1,13 @@
-package binaryTreeInorderTraversal
+package me.bossm0n5t3r.leetcode.binaryTreeInorderTraversal
 
+import me.bossm0n5t3r.leetcode.utils.TreeNode
+import me.bossm0n5t3r.leetcode.utils.TreeNodeUtil
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import utils.TreeNode
-import utils.TreeNodeUtil
 import kotlin.test.assertEquals
 
 class BinaryTreeInorderTraversalTest {
-    private val binaryTreeInorderTraversal = BinaryTreeInorderTraversal()
+    private val binaryTreeInorderTraversal = BinaryTreeInorderTraversal.Solution()
     private val treeNodeUtil = TreeNodeUtil()
 
     data class BinaryTreeInorderTraversalTestData(
